update the examples and rename to follow the "ot.solve" naming conventions
